Split Button and Label Implementation
Split the implementation of buttons and labels to be two different structs. Number labels now hold a pointer to the value they are labeling to automatically update when the value is changed. Added number label to the DrawUIElement function. Added CreateButtonElement and CreateNumberLabelElement for easier creation of a UIElement for use in an array. Added GetIndexOfKey for easier searching for a specific UIElement in an array. Changed the order of the arguments for CheckButtonInput to make more sense.