Create Button and Number Label Creators
Added functions to create buttons and number labels to enhance readability. Removed the array of UIElements in the main file, the array functionality is only needed on windows that have large amounts of UI that needs to be looped through. This array will likely be reimplemented on the settings window.