Update Scroll Selector to Animate when Pressed.
Updated the scroll selector to animate when a button is pressed and released. Added three functions to handle scroll selector input, one generic, one for pressed, and one for released. Changed the scroll selector struct to store the location of the scroll buttons to decrease the amount of calculations we are doing per frame. Added functionality for UIElements to be scroll selectors. Changed the organization of the functions in uielements to group them based on the element they correspond to.