Add OnChange for ScrollSelectors
Added an OnChange callback for scroll selectors. Added a GameSettings struct to GameState that allows for settings dependent on each other to update simultaneously via the OnChange callback for scroll selectors. The aspect ratio selector now updates the resolution selector when changed. Added functionality to the apply button in the pause menu, it will now resize the window when a new resolution is selected. Added a function to change the options in a scroll selector. Update the makefile to suppress warnings.