Added Screen Resizing Functionality
Added the ability to resize the screen between two different aspect ratios and four different resolutions. All on screen components properly reposition once the resize is complete, though currently no scaling is implemented. Implemented an initialization and resize functions to be called at game start and when a resize occurs. Moved all game objects and UI into GameState to allow for them to be repositioned when a resize occurs.